Expression with decimal Times method

+++ Wallet/Expression.cs
@@ -3,5 +3,5 @@
 public interface Expression {
   Money Reduce(Bank bank, string to);
   Expression Plus(Expression addend);
-  Expression Times(int multiplier);
+  Expression Times(decimal multiplier);
 }
